# src/Opportunity.Api/openapi/opportunity-api-v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "opportuniTY API",
    "description": "Conventions: docs/adr/0019-layering-and-api-conventions.md §2.",
    "version": "v1"
  },
  "paths": {
    "/api/v1/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "The signed-in user. Also issues the anti-forgery token cookie for later unsafe requests.",
        "operationId": "GetCurrentUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspaceId}/query-validations": {
      "post": {
        "tags": [
          "Opportunity.Api"
        ],
        "summary": "Parse query-language text and return its AST, normalized form and positioned errors.",
        "operationId": "ValidateQuery",
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryValidationResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspaceId}/jobs/{jobId}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Job status with separate committed and indexed progress.",
        "operationId": "GetJob",
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResource"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "JobCommittedProgress": {
        "required": [
          "chunksTotal",
          "chunksCommitted",
          "chunksFailed",
          "chunksCancelled",
          "chunksPending",
          "itemsApplied",
          "itemsUnchanged",
          "itemsSkippedConcurrentEdit",
          "itemsExcludedNoAccess",
          "itemsFailed"
        ],
        "type": "object",
        "properties": {
          "chunksTotal": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "chunksCommitted": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "chunksFailed": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "chunksCancelled": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "chunksPending": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "itemsApplied": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "itemsUnchanged": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "itemsSkippedConcurrentEdit": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "itemsExcludedNoAccess": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "itemsFailed": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
      "JobIndexState": {
        "enum": [
          "indexing",
          "current"
        ]
      },
      "JobIndexedProgress": {
        "required": [
          "state",
          "indexTasksTotal",
          "indexTasksApplied"
        ],
        "type": "object",
        "properties": {
          "state": {
            "$ref": "#/components/schemas/JobIndexState"
          },
          "indexTasksTotal": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "indexTasksApplied": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
      "JobResource": {
        "required": [
          "jobId",
          "workspaceId",
          "jobType",
          "status",
          "statusReason",
          "initiatedBy",
          "targetSnapshotId",
          "createdAt",
          "startedAt",
          "finishedAt",
          "committed",
          "indexed"
        ],
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid"
          },
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "jobType": {
            "$ref": "#/components/schemas/JobResourceType"
          },
          "status": {
            "$ref": "#/components/schemas/JobResourceStatus"
          },
          "statusReason": {
            "type": [
              "null",
              "string"
            ]
          },
          "initiatedBy": {
            "type": "string",
            "format": "uuid"
          },
          "targetSnapshotId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "createdAt": {
            "format": "date-time"
          },
          "startedAt": {
            "format": "date-time"
          },
          "finishedAt": {
            "format": "date-time"
          },
          "committed": {
            "$ref": "#/components/schemas/JobCommittedProgress"
          },
          "indexed": {
            "$ref": "#/components/schemas/JobIndexedProgress"
          }
        }
      },
      "JobResourceStatus": {
        "enum": [
          "created",
          "preparing",
          "running",
          "paused",
          "cancelling",
          "cancelled",
          "completed",
          "completedWithErrors",
          "failed"
        ]
      },
      "JobResourceType": {
        "enum": [
          "import",
          "bulkCoding",
          "relationshipFixup",
          "reindex",
          "export",
          "production",
          "render"
        ]
      },
      "JsonElement": { },
      "MeResponse": {
        "required": [
          "userId",
          "displayName",
          "email",
          "groups",
          "mfa",
          "sessionExpiresAt"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "groups": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "mfa": {
            "type": "boolean"
          },
          "sessionExpiresAt": {
            "format": "date-time"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "QueryValidationDiagnostic": {
        "required": [
          "code",
          "message",
          "span",
          "expected"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "span": {
            "$ref": "#/components/schemas/TextSpan"
          },
          "expected": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "QueryValidationRequest": {
        "required": [
          "query"
        ],
        "type": "object",
        "properties": {
          "query": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "QueryValidationResult": {
        "required": [
          "valid",
          "astVersion",
          "errors",
          "warnings"
        ],
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean"
          },
          "astVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "ast": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JsonElement"
              }
            ]
          },
          "normalized": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueryValidationDiagnostic"
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueryValidationDiagnostic"
            }
          }
        }
      },
      "TextSpan": {
        "required": [
          "start",
          "end"
        ],
        "type": "object",
        "properties": {
          "start": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "end": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Authentication"
    },
    {
      "name": "Opportunity.Api"
    },
    {
      "name": "Jobs"
    }
  ]
}