# src/Opportunity.Web/src/app/core/api/generated/models/query-validation-diagnostic.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { TextSpan } from '../models/text-span';
export interface QueryValidationDiagnostic {
  code: string;
  expected: Array<string>;
  message: string;
  span: TextSpan;
}