# src/Opportunity.Web/src/app/core/api/generated/models/query-validation-result.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { JsonElement } from '../models/json-element';
import { QueryValidationDiagnostic } from '../models/query-validation-diagnostic';
export interface QueryValidationResult {
  ast?: null | JsonElement;
  astVersion: number | string;
  errors: Array<QueryValidationDiagnostic>;
  normalized?: string | null;
  valid: boolean;
  warnings: Array<QueryValidationDiagnostic>;
}